# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/subscriptions": {
      "get": {
        "description": "Get all subscriptions optionally filtered by user_id and service_name",
        "produces": [
          "application/json"
        ],
        "tags": [
          "subscriptions"
        ],
        "summary": "List subscriptions",
        "parameters": [
          {
            "type": "string",
            "description": "Filter by user UUID",
            "name": "user_id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Filter by service name",
            "name": "service_name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/model.Subscription"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a subscription with service name, price, user ID, start and optional end dates",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "subscriptions"
        ],
        "summary": "Create a new subscription",
        "parameters": [
          {
            "description": "Subscription request body",
            "name": "subscription",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.SubscriptionReq"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/model.Subscription"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/subscriptions/total": {
      "get": {
        "description": "Calculate total cost for a user and optional service within a date range (from, to in MM-YYYY format)",
        "produces": [
          "application/json"
        ],
        "tags": [
          "subscriptions"
        ],
        "summary": "Calculate total subscription cost",
        "parameters": [
          {
            "type": "string",
            "description": "User UUID",
            "name": "user_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "Service name",
            "name": "service_name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "Start period (MM-YYYY)",
            "name": "from",
            "in": "query"
          },
          {
            "type": "string",
            "description": "End period (MM-YYYY)",
            "name": "to",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "integer"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "description": "Get details of a subscription by its UUID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "subscriptions"
        ],
        "summary": "Get subscription by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Subscription ID (UUID)",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.Subscription"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "put": {
        "description": "Update subscription fields by subscription UUID",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "subscriptions"
        ],
        "summary": "Update subscription by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Subscription ID (UUID)",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Updated subscription data",
            "name": "subscription",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/model.SubscriptionReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/model.Subscription"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete subscription by UUID",
        "tags": [
          "subscriptions"
        ],
        "summary": "Delete subscription by ID",
        "parameters": [
          {
            "type": "string",
            "description": "Subscription ID (UUID)",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "model.Subscription": {
      "type": "object",
      "properties": {
        "end_date": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "price": {
          "type": "integer"
        },
        "service_name": {
          "type": "string"
        },
        "start_date": {
          "description": "формат \"07-2025\"",
          "type": "string"
        },
        "user_id": {
          "type": "string"
        }
      }
    },
    "model.SubscriptionReq": {
      "type": "object",
      "required": [
        "price",
        "service_name",
        "start_date",
        "user_id"
      ],
      "properties": {
        "end_date": {
          "description": "EndDate optional subscription end date in MM-YYYY format",
          "type": "string"
        },
        "price": {
          "description": "Price subscription price, must be positive\nrequired: true",
          "type": "integer"
        },
        "service_name": {
          "description": "ServiceName is the name of the subscription service\nrequired: true",
          "type": "string"
        },
        "start_date": {
          "description": "StartDate subscription start date in MM-YYYY format\nrequired: true",
          "type": "string",
          "example": "07-2025"
        },
        "user_id": {
          "description": "UserID owner of the subscription (UUID)\nrequired: true",
          "type": "string"
        }
      }
    }
  }
}